Introduce view controller test fixture
